% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-level.R
\name{power2L}
\alias{power2L}
\title{power2L}
\usage{
power2L(n, J, delta, rhoL2, numcovL1 = 0, numcovL2 = 0, R2L1 = 0,
  R2L2 = 0, modL1 = FALSE, modL2 = FALSE, alpha = 0.05)
}
\arguments{
\item{n}{Number of units in level 1 (individual level).}

\item{J}{Number of units in level 2 (cluster level)}

\item{delta}{Standardized effect size.}

\item{rhoL2}{Intraclass correlation (ICC) at level 2.}

\item{numcovL1}{Number of covariate(s) at level 1.}

\item{numcovL2}{Number of covariate(s) at level 2.}

\item{R2L1}{R-squared at level 1.}

\item{R2L2}{R-squared at level 2.}

\item{modL1}{Existence of moderator at level 1.}

\item{modL2}{Existence of moderator at level 2.}

\item{alpha}{Probability of type I error.}
}
\description{
This function returns the power for a two-level cluster randomized trials designs.
}
\examples{
power2L(n=100,J=40,delta=0.2,rhoL2=0.3,numcovL1=1,numcovL2=0,R2L1=0.2,R2L2=0,modL1=TRUE,modL2=FALSE,alpha=0.05)
}
\seealso{
\code{\link{PowerUpR}}
}
