% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three-level.R
\name{power3L}
\alias{power3L}
\title{power3L}
\usage{
power3L(n, J, K, delta, rhoL2, rhoL3, numcovL1 = 0, numcovL2 = 0,
  numcovL3 = 0, R2L1 = 0, R2L2 = 0, R2L3 = 0, modL1 = FALSE,
  modL2 = FALSE, modL3 = FALSE, alpha = 0.05)
}
\arguments{
\item{n}{Number of subjects in level 1 (individual level).}

\item{J}{Number of units in level 2 (subcluster level).}

\item{K}{Number of units in level 3 (cluster level).}

\item{delta}{Standardized effect size.}

\item{rhoL2}{Intraclass correlation (ICC) at level 2.}

\item{rhoL3}{Intraclass correlation (ICC) at level 3.}

\item{numcovL1}{Number of covariate(s) at level 1.}

\item{numcovL2}{Number of covariate(s) at level 2.}

\item{numcovL3}{Number of covariate(s) at level 3.}

\item{R2L1}{R-squared at level 1.}

\item{R2L2}{R-squared at level 2.}

\item{R2L3}{R-squared at level 3.}

\item{modL1}{Existence of moderator at level 1.}

\item{modL2}{Existence of moderator at level 2.}

\item{modL3}{Existence of moderator at level 3.}

\item{alpha}{Probability of type I error.}
}
\description{
This function returns the power for a three-level cluster randomized trials designs.
}
\examples{
power3L(n=100,J=40,K=40,delta=0.2,rhoL2=0.3,rhoL3=0.2,numcovL1=0,numcovL2=0,numcovL3=0,R2L1=0,R2L2=0,modL1=FALSE,modL2=FALSE,modL3=FALSE,alpha=0.05)

}
\seealso{
\code{\link{PowerUpR}}
}
