% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisite.R
\name{powerMS}
\alias{powerMS}
\title{powerMS}
\usage{
powerMS(n, J, delta, sigmad, rhoL2, numcovL1 = 0, R2L1 = 0,
  randeff = TRUE, alpha = 0.05)
}
\arguments{
\item{n}{Number of subjects in level 1 (individual level).}

\item{J}{Number of units in level 2 (site level).}

\item{delta}{Standardized effect size.}

\item{sigmad}{Variance in treatment effects across sites.}

\item{rhoL2}{Intraclass correlation (ICC) at level 2.}

\item{numcovL1}{Number of covariate(s) at level 1.}

\item{R2L1}{R-squared at level 1.}

\item{randeff}{Random site effects is assumed.}

\item{alpha}{Probability of type I error.}
}
\description{
This function returns the power for a multi-site (blocked) trials designs.
}
\examples{
# All arguments are specified
powerMS(n=100,J=40,delta=0.2,sigmad=0.2,rhoL2=0.3,numcovL1=1,R2L1=0.2,randeff=FALSE,alpha=0.05)
# Unspecified arguments take default values
powerMS(n=100,J=40,delta=0.2,sigmad=0.2,rhoL2=0.3)
}
\seealso{
\code{\link{PowerUpR}}
}
